pyproject.toml
[project]
name = "mameformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "numpy",
    "tensorboard",
    "onnx>=1.20.0",
    "onnxscript>=0.5.6",
    "hydra-core>=1.3.2",
    "accelerate>=0.34.2",
    "mlflow>=2.17.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.uv.pip]
torch-backend = "auto"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # 行長は formatter に任せる
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

